New rx_agc implementation 1/2
New AGC implementation with less complexity.
CPU load does not vary depending on input.
New controls, that should be easy to understand:
Maximum gain - maximum AGC gain in auto mode.
Target level - a level, to normalize the signal to in dB.
Attack - attack time in ms. Time to go from max gain to 0 dB (worst
case). Defines the size of the delay buffer and, as
a result, sets audio delay, introduced by the block.
Decay - decay time in ms. Time to go from 0 dB gain to maximum gain
(worst case).
Hang - hang time in ms. Time to wait before increasing the gain after
the peak.

GUI and setting are updated.

The next step is to put the AGC block after the demodulator to make it
usable in FM mode and to remove useless audio_gainX blocks.