plotter: consistently show signal stength labels
Depending on the vertical position and zoom/scale, the top most signal
strength label was sometimes not being drawn despite the associated
line being drawn and there being plenty of space to draw the label. This
was due to a mismatch of <= being used for counting the lines, but <
being used for counting the labels.

The new logic makes the label indexing consistent with the line drawing
indexing, while adding a check to make sure the label is not cut off
vertically.